Widen tile values to ushort; extend TileBitPacker to 10 bits
Tile values (FrameTileMap, TileBitPacker, TileClassification.PaletteIndex,
the decoder's classification arrays) become ushort, and TileBitPacker
now packs 1-10 bits, so a frame can carry 9- or 10-bit tiles for the
512/1024-colour palettes. Behaviour-neutral at 256/8-bit: values 0-255
round-trip identically, so rendered frames and the golden suite are
unchanged. Adds packer round-trip coverage at 9 and 10 bits.

+++ FluxCore.Tests/Framing/FrameCodecDepthTests.cs
@@ -14,9 +14,9 @@ public class FrameCodecDepthTests
         return data;
     }
 
-    private static byte[] DataTileValues(FrameTileMap map)
+    private static ushort[] DataTileValues(FrameTileMap map)
     {
-        var values = new byte[FrameFormat.DataTiles.Count];
+        var values = new ushort[FrameFormat.DataTiles.Count];
         for (int t = 0; t < values.Length; t++)
         {
             var (x, y) = FrameFormat.DataTiles[t];
